Fixed Purchase Logic, refactored controllers with viewControllerService, changed purchaseCell to view

+++ SenSoyleAvrupa/Model/PurchaseModel.swift
@@ -9,6 +9,6 @@ import Foundation
 
 struct PurchaseModel {
     var coin: Int
-    var price: Int
+    var price: Float
     var handler: (() -> Void)
 }
